+++ apps/core/__init__.py
@@ -12,3 +12,5 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this file.  If not, see <https://www.gnu.org/licenses/>.
+
+"""The top-level module for core application."""